+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Use total visible time as duration in `neat_message` and `neat_sentence`
 - Add missing killtarget to `neat_message`
 - Add `neat_sentence`
+- Fix mod filter-materials not handling prefixes
 
 # v1.7.0
 - Remove erroneous textures lump size limit